+++ core/filter.h
@@ -38,6 +38,8 @@ namespace oidn {
     virtual void setImage(const std::string& name, const Image& data) = 0;
     virtual void set1i(const std::string& name, int value) = 0;
     virtual int get1i(const std::string& name) = 0;
+    virtual void set1f(const std::string& name, float value) = 0;
+    virtual float get1f(const std::string& name) = 0;
 
     void setProgressMonitorFunction(ProgressMonitorFunction func, void* userPtr);